Extracted select_columns_map to module & namespaced classes
